pyproject.toml
[project]
name = "MeshSupportList"
version = "1.0"
dependencies = [
    "requests",
    "python-dotenv",
    "slack-sdk>=3.12,<4",
    "pytz",
    "python-dateutil"
]

[project.optional-dependencies]
test = [
    "pytest >= 6",
    "coverage",
    "freezegun"
]

[project.scripts]
mesh-support-report = "mesh_support_report_generator:main.main"